Refactor:드랍다운 파일 재배치, 파일명 컴포넌트명 일치하게 수정
